Ukrainian translation
This is a Ukrainian translation. Please comment the line in the source code that says the face is well positioned, otherwise the translation will not work. When we published it on the channel, the translator commented on the line in the source code, so the translation works there. In other words, make this line available for translation so that the translation works correctly.